IDVA6-2392 Logic modified to replace hardcoded value with a language key
