+++ locales/cy/common.json
@@ -32,6 +32,7 @@
     "next": "Nesaf",
     "next_page": "Tudalen nesaf",
     "noscript_message": "Rydym yn defnyddio cwcis i wneud i’n gwasanaethau weithio a chasglu gwybodaeth ddadansoddeg. I dderbyn neu wrthod cwcis dadansoddeg, trowch JavaScript ymlaen yn ngosodiadau eich porwr ac ail-lwytho’r dudalen hon.",
+    "not_provided": "Heb ei ddarparu",
     "open_government_licence": "Drwydded Llywodraeth Agored v3.0",
     "open_government_licence_url": "https://www.nationalarchives.gov.uk/doc/open-government-licence-cymraeg/version/3/",
     "page": "Tudalen",
