+++ locales/en/common.json
@@ -32,6 +32,7 @@
     "next": "Next",
     "next_page": "Next page",
     "noscript_message": "We use cookies to make our services work and collect analytics information. To accept or reject analytics cookies, turn on JavaScript in your browser settings and reload this page.",
+    "not_provided": "Not Provided",
     "open_government_licence": "Open Government Licence v3.0",
     "open_government_licence_url": "https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/",
     "page": "Page",
