+++ src/lib/constants.ts
@@ -96,12 +96,12 @@ export const ITEMS_PER_PAGE_DEFAULT = 15;
 export const UNCAUGHT_EXCEPTION = "uncaughtException";
 export const UNHANDLED_REJECTION = "unhandledRejection";
 export const NOT_PROVIDED = "Not Provided";
-export const NOT_PROVIDED_CY = "Heb ei ddarparu";
 export const FEATURE_FLAG_SHOW_FILE_AS_AUTHORISED_AGENT = "FEATURE_FLAG_SHOW_FILE_AS_AUTHORISED_AGENT";
 export const FEATURE_FLAG_SHOW_UPDATE_AUTHORISED_AGENT_DETAILS = "FEATURE_FLAG_SHOW_UPDATE_AUTHORISED_AGENT_DETAILS";
 export const FEATURE_FLAG_SHOW_CLOSE_AUTHORISED_AGENT = "FEATURE_FLAG_SHOW_CLOSE_AUTHORISED_AGENT";
 export const FEATURE_FLAG_SHOW_TELL_US_YOUVE_VERIFIED_A_PERSONS_IDENTITY = "FEATURE_FLAG_SHOW_TELL_US_YOUVE_VERIFIED_A_PERSONS_IDENTITY";
 export const REFRESH_TOKEN_GRANT_TYPE = "refresh_token";
+export const LANG_KEY_FOR_NOT_PROVIDED = "not_provided";
 
 // URL Base Paths and Helpers
 export const HEALTHCHECK = "/healthcheck";
